+++ src/main/java/com/revomatico/play/javaship2020/impl/CristianOMovieImporter.java
@@ -1,8 +1,11 @@
-package com.revomatico.play.javaship2020;
+package com.revomatico.play.javaship2020.impl;
 
 import com.opencsv.CSVReader;
 import com.opencsv.CSVReaderBuilder;
 import com.opencsv.exceptions.CsvValidationException;
+import com.revomatico.play.javaship2020.Movie;
+import com.revomatico.play.javaship2020.MovieImporter;
+import com.revomatico.play.javaship2020.PopcornApp;
 
 import java.io.FileReader;
 import java.io.IOException;
