+++ src/main/java/com/revomatico/play/javaship2020/impl/RobertMovieImporter.java
@@ -1,8 +1,11 @@
-package com.revomatico.play.javaship2020;
+package com.revomatico.play.javaship2020.impl;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
+
+import com.revomatico.play.javaship2020.Movie;
+import com.revomatico.play.javaship2020.MovieImporter;
 
 import java.io.File;
 import java.io.FileNotFoundException;
