+++ src/test/java/com/revomatico/play/javaship2020/impl/LucianTest.java
@@ -1,8 +1,9 @@
-package com.revomatico.play.javaship2020;
+package com.revomatico.play.javaship2020.impl;
 
 import java.io.FileNotFoundException;
 import java.text.ParseException;
 
+import com.revomatico.play.javaship2020.impl.LucianMovieImporter;
 import org.junit.Test;
 
 public class LucianTest {
